pyproject.toml
[project]
name = "isomcom"
version = "0.1.0"
description = "LINE Repair Management System"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "psycopg2-binary",
    "python-dotenv",
    "line-bot-sdk",
    "pydantic[email]",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "bcrypt==4.0.1",
    "python-multipart",
    "aiofiles>=25.1.0",
    "jinja2>=3.1.6",
    "pillow>=12.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "pylint>=4.0.5",
    "ruff>=0.15.5",
]
